requirements.txt
spotipy>=2.23.0
pyperclip>=1.8.2
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0